# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/requisitions": {
      "get": {
        "operationId": "RequisitionsController_list",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "The company the person acts in (a company UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Requisition"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Newest first: the person's own requisitions; for an approver also the submitted ones routed\nto approvers and those they decided; for a buyer the approved ones (to order them); for an\nadmin every one of the company.",
        "tags": [
          "Requisitions"
        ]
      },
      "post": {
        "operationId": "RequisitionsController_create",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "The company the person acts in (a company UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRequisitionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requisition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Saves a new draft as the person asking.",
        "tags": [
          "Requisitions"
        ]
      }
    },
    "/requisitions/{id}": {
      "get": {
        "operationId": "RequisitionsController_get",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "The company the person acts in (a company UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requisition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Requisitions"
        ]
      },
      "put": {
        "operationId": "RequisitionsController_update",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "The company the person acts in (a company UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRequisitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requisition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replaces a draft's cost center, justification and lines.",
        "tags": [
          "Requisitions"
        ]
      }
    },
    "/requisitions/{id}/submit": {
      "post": {
        "operationId": "RequisitionsController_submit",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "The company the person acts in (a company UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requisition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Draft to submitted, refused unless it has a cost center, a justification and a line. Under\nthe company's lowest approval threshold it is approved at once, with the reason in\ndecisionNote. A company without approval rules leaves it to an admin.",
        "tags": [
          "Requisitions"
        ]
      }
    },
    "/requisitions/{id}/cancel": {
      "post": {
        "operationId": "RequisitionsController_cancel",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "The company the person acts in (a company UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requisition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Draft or submitted to cancelled. Cancelled is final.",
        "tags": [
          "Requisitions"
        ]
      }
    },
    "/requisitions/{id}/approve": {
      "post": {
        "operationId": "RequisitionsController_approve",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "The company the person acts in (a company UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequisitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requisition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submitted to approved, by an approver or admin the requisition's route admits. Never by the\nrequester. The comment is optional and shown to the requester.",
        "tags": [
          "Requisitions"
        ]
      }
    },
    "/requisitions/{id}/reject": {
      "post": {
        "operationId": "RequisitionsController_reject",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "The company the person acts in (a company UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequisitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Requisition"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submitted to rejected, by the same deciders. The reason is required and shown to the requester.",
        "tags": [
          "Requisitions"
        ]
      }
    },
...
    "/purchase-orders": {
      "get": {
        "operationId": "PurchaseOrdersController_list",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "The company the person acts in (a company UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PurchaseOrder"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Newest first. Buyers and admins get every order of the company; a requester the orders of\ntheir own requisitions; an approver those of the requisitions they decided.",
        "tags": [
          "PurchaseOrders"
        ]
      },
      "post": {
        "operationId": "PurchaseOrdersController_create",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "The company the person acts in (a company UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrder"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buyers and admins only. Refused unless the requisition is approved and has no order yet,\nthe supplier is active, and every item belongs to that supplier. A requisition converts\nonce; a second attempt is a 409.",
        "tags": [
          "PurchaseOrders"
        ]
      }
    },
    "/purchase-orders/{id}": {
      "get": {
        "operationId": "PurchaseOrdersController_get",
        "parameters": [
          {
            "name": "x-company-id",
            "in": "header",
            "description": "The company the person acts in (a company UUID).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrder"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "PurchaseOrders"
        ]
      }
    }
  },
  "info": {
    "title": "Procurely API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RequisitionStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "SUBMITTED",
          "APPROVED",
          "REJECTED",
          "CANCELLED"
        ]
      },
      "ApprovalRoute": {
        "type": "string",
        "enum": [
          "APPROVER",
          "ADMIN",
          "UNDER_THRESHOLD",
          "NO_RULES"
        ],
        "description": "Null while a draft, and on a draft that was cancelled."
      },
      "RequisitionLine": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "catalogItemId": {
            "type": "string"
          },
          "catalogItemName": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "description": "A whole number, 1 or more."
          },
          "unitPriceMinor": {
            "type": "number",
            "description": "Copied from the catalog item when the requisition was saved, in integer minor units."
          },
          "amountMinor": {
            "type": "number",
            "description": "quantity times unitPriceMinor."
          }
        },
        "required": [
          "id",
          "catalogItemId",
          "catalogItemName",
          "quantity",
          "unitPriceMinor",
          "amountMinor"
        ]
      },
      "RequisitionAction": {
        "type": "string",
        "enum": [
          "edit",
          "submit",
          "cancel",
          "approve",
          "reject"
        ],
        "description": "What the signed-in person may do to it now. Submitting a draft can still be refused when\nit is incomplete."
      },
      "Requisition": {
        "type": "object",
        "properties": {
          "costCenterId": {
            "type": "string",
            "nullable": true,
            "description": "Null while a draft has no cost center yet."
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RequisitionStatus"
              }
            ]
          },
          "approvalRoute": {
            "nullable": true,
            "description": "Null while a draft, and on a draft that was cancelled.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ApprovalRoute"
              }
            ]
          },
          "decisionNote": {
            "type": "string",
            "nullable": true,
            "description": "For the requester: the approver's comment, the rejection reason, or why the requisition was\napproved automatically. Null when there is none."
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequisitionLine"
            }
          },
          "actions": {
            "type": "array",
            "description": "What the signed-in person may do to it now. Submitting a draft can still be refused when\nit is incomplete.",
            "items": {
              "$ref": "#/components/schemas/RequisitionAction"
            }
          },
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "requesterPersonId": {
            "type": "string"
          },
          "requesterName": {
            "type": "string"
          },
          "justification": {
            "type": "string"
          },
          "totalMinor": {
            "type": "number",
            "description": "Sum of the line amounts in integer minor units of the company currency."
          }
        },
        "required": [
          "costCenterId",
          "status",
          "approvalRoute",
          "decisionNote",
          "lines",
          "actions",
          "id",
          "companyId",
          "requesterPersonId",
          "requesterName",
          "justification",
          "totalMinor"
        ]
      },
      "RequisitionLineInput": {
        "type": "object",
        "properties": {
          "catalogItemId": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "description": "A whole number, 1 or more."
          }
        },
        "required": [
          "catalogItemId",
          "quantity"
        ]
      },
      "SaveRequisitionRequest": {
        "type": "object",
        "properties": {
          "costCenterId": {
            "type": "string",
            "nullable": true
          },
          "lines": {
            "description": "Items of active suppliers. The unit prices are taken from the catalog.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequisitionLineInput"
            }
          },
          "justification": {
            "type": "string",
            "description": "May be empty in a draft."
          }
        },
        "required": [
          "costCenterId",
          "lines",
          "justification"
        ]
      },
      "ApproveRequisitionRequest": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "description": "Shown to the requester. Optional; blank counts as none."
          }
        }
      },
      "RejectRequisitionRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Shown to the requester. Required, not blank."
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "PurchaseOrderLine": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "catalogItemId": {
            "type": "string"
          },
          "catalogItemName": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "description": "A whole number, 1 or more."
          },
          "unitPriceMinor": {
            "type": "number",
            "description": "Set by the buyer when the order was made, in integer minor units."
          },
          "amountMinor": {
            "type": "number",
            "description": "quantity times unitPriceMinor."
          }
        },
        "required": [
          "id",
          "catalogItemId",
          "catalogItemName",
          "quantity",
          "unitPriceMinor",
          "amountMinor"
        ]
      },
      "PurchaseOrder": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "description": "ISO 8601 timestamp.",
            "format": "date-time"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseOrderLine"
            }
          },
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "requisitionId": {
            "type": "string"
          },
          "requisitionJustification": {
            "type": "string",
            "description": "The requisition's justification, to recognise it by."
          },
          "supplierId": {
            "type": "string"
          },
          "supplierName": {
            "type": "string"
          },
          "createdByPersonId": {
            "type": "string"
          },
          "createdByName": {
            "type": "string"
          },
          "totalMinor": {
            "type": "number",
            "description": "Sum of the line amounts in integer minor units of the company currency."
          }
        },
        "required": [
          "createdAt",
          "lines",
          "id",
          "companyId",
          "requisitionId",
          "requisitionJustification",
          "supplierId",
          "supplierName",
          "createdByPersonId",
          "createdByName",
          "totalMinor"
        ]
      },
      "PurchaseOrderLineInput": {
        "type": "object",
        "properties": {
          "catalogItemId": {
            "type": "string",
            "description": "An item of the chosen supplier. Each item may appear once."
          },
          "quantity": {
            "type": "number",
            "description": "A whole number, 1 or more."
          },
          "unitPriceMinor": {
            "type": "number",
            "description": "Integer minor units of the company currency, zero or more."
          }
        },
        "required": [
          "catalogItemId",
          "quantity",
          "unitPriceMinor"
        ]
      },
      "CreatePurchaseOrderRequest": {
        "type": "object",
        "properties": {
          "lines": {
            "description": "At least one line.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseOrderLineInput"
            }
          },
          "requisitionId": {
            "type": "string",
            "description": "An approved requisition of the company that has no purchase order yet."
          },
          "supplierId": {
            "type": "string",
            "description": "An active supplier of the company."
          }
        },
        "required": [
          "lines",
          "requisitionId",
          "supplierId"
        ]
      }
    }
  }
}